requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
